adding configurations

+++ DollarsBankWebApp/src/com/cognixia/model/Transactions.java
@@ -6,13 +6,13 @@ public class Transactions {
 	
 	private static int count = 1;
 	private int id;
-	private int customerId;
+	private Long customerId;
 	private String type;
 	private double amount;
 	private Date date;
 	
 	
-	public Transactions(int customerId, String type, double amount, Date date) {
+	public Transactions(Long customerId, String type, double amount, Date date) {
 		super();
 		this.id = count++;
 		this.customerId = customerId;
@@ -32,12 +32,12 @@ public class Transactions {
 	}
 
 
-	public int getCustomerId() {
+	public Long getCustomerId() {
 		return customerId;
 	}
 
 
-	public void setCustomerId(int customerId) {
+	public void setCustomerId(Long customerId) {
 		this.customerId = customerId;
 	}
